分离判断逻辑
